+++ article/src/main/java/com/iag/bbscloud/article/model/ArticleStatistic.java
@@ -21,12 +21,12 @@ public class ArticleStatistic {
 
     @Id
     @Column(nullable = false, unique = true)
-    private BigInteger                          aid;
+    private Long                                aid;
     @Column
-    private BigInteger                          uid;
+    private Long                                uid;
     // 板块ID
     @Column
-    private BigInteger                          bid;
+    private Long                                bid;
     @Column
     private Integer                             viewCount;
     @Column
