+++ article/src/main/java/com/iag/bbscloud/article/model/repository/ArticleRepository.java
@@ -9,6 +9,6 @@ import java.math.BigInteger;
  * @Author xueshan.wei@mljr.com
  * @Date 2017/9/26 下午3:37
  */
-public interface ArticleRepository extends CrudRepository<Article, BigInteger> {
+public interface ArticleRepository extends CrudRepository<Article, Long> {
 
 }
